[Amit] Move CalendarEventAlertUIModel and ViewAlertType into model package
