+++ app/src/main/java/com/mymeetings/android/model/CalendarEventAlert.kt
@@ -1,7 +1,5 @@
 package com.mymeetings.android.model
 
-import com.mymeetings.android.ViewAlertType
-
 data class CalendarEventAlert(
     val startTime: Long,
     val reminderTime: Long
