+++ app/src/main/java/com/mymeetings/android/model/ViewAlertType.kt
@@ -1,4 +1,4 @@
-package com.mymeetings.android
+package com.mymeetings.android.model
 
 enum class ViewAlertType {
     RUNNING,
