+++ app/src/main/java/com/mymeetings/android/view/widgets/CalendarEventWidgetViewProvider.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import android.view.View
 import android.widget.RemoteViews
 import com.mymeetings.android.R
-import com.mymeetings.android.ViewAlertType
-import com.mymeetings.android.view.uiModels.CalendarEventAlertUIModel
+import com.mymeetings.android.model.ViewAlertType
+import com.mymeetings.android.model.CalendarEventAlertUIModel
 
 class CalendarEventWidgetViewProvider(
     private val context: Context
